requirements.txt
flask==3.0.3
pyyaml==6.0.2
gunicorn==22.0.0
pytest==8.2.0
